+++ src/background/utils/rect.rs
@@ -1,7 +1,7 @@
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 use windows::Win32::Foundation::RECT;
 
-#[derive(Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug, Clone, Default)]
 pub struct Rect {
     pub left: i32,
     pub top: i32,